Introduce `BaseNodeModifier.java`
