+++ compiler/ballerina-treegen/src/main/java/io/ballerinalang/compiler/internal/treegen/TreeGen.java
@@ -25,6 +25,7 @@ import io.ballerinalang.compiler.internal.treegen.targets.SourceText;
 import io.ballerinalang.compiler.internal.treegen.targets.Target;
 import io.ballerinalang.compiler.internal.treegen.targets.node.ExternalNodeTarget;
 import io.ballerinalang.compiler.internal.treegen.targets.node.InternalNodeTarget;
+import io.ballerinalang.compiler.internal.treegen.targets.nodevisitor.BaseNodeModifierTarget;
 import io.ballerinalang.compiler.internal.treegen.targets.nodevisitor.ExternalNodeVisitorTarget;
 import io.ballerinalang.compiler.internal.treegen.targets.nodevisitor.InternalNodeTransformerTarget;
 import io.ballerinalang.compiler.internal.treegen.targets.nodevisitor.InternalNodeVisitorTarget;
@@ -87,6 +88,7 @@ public final class TreeGen {
         targetList.add(new TreeModifierTarget(config));
         targetList.add(new InternalTreeModifierTarget(config));
         targetList.add(new NodeFactoryTarget(config));
+        targetList.add(new BaseNodeModifierTarget(config));
         return targetList;
     }
