+++ compiler/ballerina-treegen/src/main/java/io/ballerinalang/compiler/internal/treegen/targets/nodevisitor/AbstractNodeVisitorTarget.java
@@ -48,6 +48,7 @@ public abstract class AbstractNodeVisitorTarget extends Target {
     static final String EXTERNAL_NODE_FACTORY_CN = "NodeFactory";
     static final String EXTERNAL_BASE_NODE_FACTORY_CN = "AbstractNodeFactory";
     public static final String EXTERNAL_NODE_VISITOR_CN = "NodeVisitor";
+    public static final String BASE_NODE_MODIFIER_CN = "BaseNodeModifier";
 
     AbstractNodeVisitorTarget(TreeGenConfig config) {
         super(config);
