+++ compiler/ballerina-treegen/src/main/resources/external_base_node_modifier_template.mustache
@@ -1,0 +1,153 @@
+/*
+ *  Copyright (c) 2025, WSO2 LLC. (http://www.wso2.com).
+ *
+ *  WSO2 LLC. licenses this file to you under the Apache License,
+ *  Version 2.0 (the "License"); you may not use this file except
+ *  in compliance with the License.
+ *  You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *  Unless required by applicable law or agreed to in writing,
+ *  software distributed under the License is distributed on an
+ *  "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ *  KIND, either express or implied. See the License for the
+ *  specific language governing permissions and limitations
+ *  under the License.
+ */
+package {{packageName}};
+
+{{#imports}}
+import {{name}};
+{{/imports}}
+
+import java.util.function.Function;
+
+/**
+ * Produces a new tree by doing a depth-first traversal of the tree.
+ * New tree nodes can be replaced with any subclass of their base class.
+ *
+ * This is a generated class.
+ *
+ * @since 2201.13.0
+ */
+public abstract class {{className}} extends {{superClassName}}<Node> {
+    {{#nodes}}
+    {{^isAbstract}}
+
+    @Override
+    public {{superClassName}} transform(
+            {{externalClassName}} {{camelCaseExternalClassName}}) {
+        {{#fields}}
+        {{#isList}}{{listClassName}}<{{fieldClassName}}>{{/isList}}{{^isList}}{{fieldClassName}}{{/isList}} {{fieldName}} =
+                modify{{#isList}}{{listClassName}}{{/isList}}{{^isList}}{{#isToken}}Token{{/isToken}}{{#isNode}}Node{{/isNode}}{{/isList}}({{camelCaseExternalClassName}}.{{fieldName}}(){{#isOptional}}.orElse(null){{/isOptional}});
+        {{/fields}}
+        return {{camelCaseExternalClassName}}.modify(
+                {{^syntaxKind}}{{camelCaseExternalClassName}}.kind(), &#10;                {{/syntaxKind}}{{#fields}}{{fieldName}}{{^isLast}}, &#10;                {{/isLast}}{{/fields}});
+    }
+    {{/isAbstract}}
+    {{/nodes}}
+
+    // Tokens
+
+    @Override
+    public Token transform(Token token) {
+        return token;
+    }
+
+    @Override
+    public IdentifierToken transform(IdentifierToken identifier) {
+        return identifier;
+    }
+
+    @Override
+    protected Node transformSyntaxNode(Node node) {
+        return node;
+    }
+
+    protected <T extends Node> NodeList<T> modifyNodeList(NodeList<T> nodeList) {
+        return modifyGenericNodeList(nodeList, NodeList::new);
+    }
+
+    protected <T extends Node> SeparatedNodeList<T> modifySeparatedNodeList(SeparatedNodeList<T> nodeList) {
+        Function<NonTerminalNode, SeparatedNodeList> nodeListCreator = SeparatedNodeList::new;
+        if (nodeList.isEmpty()) {
+            return nodeList;
+        }
+
+        boolean nodeModified = false;
+        STNode[] newSTNodes = new STNode[nodeList.internalListNode.size()];
+
+        for (int index = 0; index < nodeList.size(); index++) {
+            T oldNode = nodeList.get(index);
+            T newNode = modifyNode(oldNode);
+            if (oldNode != newNode) {
+                nodeModified = true;
+            }
+
+            newSTNodes[2 * index] = newNode.internalNode();
+            if (index == nodeList.size() - 1) {
+                break;
+            }
+
+            Token oldSeperator = nodeList.getSeparator(index);
+            Token newSeperator = modifyToken(oldSeperator);
+
+            if (oldSeperator != newSeperator) {
+                nodeModified = true;
+            }
+
+            newSTNodes[(2 * index) + 1] = newSeperator.internalNode();
+        }
+
+        if (!nodeModified) {
+            return nodeList;
+        }
+
+        STNode stNodeList = STNodeFactory.createNodeList(java.util.Arrays.asList(newSTNodes));
+        return nodeListCreator.apply(stNodeList.createUnlinkedFacade());
+    }
+
+    private <T extends Node, N extends NodeList<T>> N modifyGenericNodeList(
+            N nodeList,
+            Function<NonTerminalNode, N> nodeListCreator) {
+        if (nodeList.isEmpty()) {
+            return nodeList;
+        }
+
+        boolean nodeModified = false;
+        STNode[] newSTNodes = new STNode[nodeList.size()];
+        for (int index = 0; index < nodeList.size(); index++) {
+            T oldNode = nodeList.get(index);
+            T newNode = modifyNode(oldNode);
+            if (oldNode != newNode) {
+                nodeModified = true;
+            }
+            newSTNodes[index] = newNode.internalNode();
+        }
+
+        if (!nodeModified) {
+            return nodeList;
+        }
+
+        STNode stNodeList = STNodeFactory.createNodeList(java.util.Arrays.asList(newSTNodes));
+        return nodeListCreator.apply(stNodeList.createUnlinkedFacade());
+    }
+
+    protected <T extends Token> T modifyToken(T token) {
+        if (token == null) {
+            return null;
+        }
+        // TODO
+        return (T) token.apply(this);
+    }
+
+    protected <T extends Node> T modifyNode(T node) {
+        if (node == null) {
+            return null;
+        }
+        // TODO
+        return (T) node.apply(this);
+    }
+}
+
