+++ compiler/ballerina-treegen/src/main/resources/external_tree_modifier_template.mustache
@@ -21,132 +21,23 @@ package {{packageName}};
 import {{name}};
 {{/imports}}
 
-import java.util.function.Function;
-
 /**
  * Produces a new tree by doing a depth-first traversal of the tree.
+ * New tree nodes can only be replaced with nodes of the same kind.
  *
  * This is a generated class.
  *
  * @since 2.0.0
  */
-public abstract class {{className}} extends {{superClassName}}<Node> {
+public abstract class {{className}} extends {{superClassName}} {
     {{#nodes}}
     {{^isAbstract}}
 
     @Override
     public {{externalClassName}} transform(
             {{externalClassName}} {{camelCaseExternalClassName}}) {
-        {{#fields}}
-        {{#isList}}{{listClassName}}<{{fieldClassName}}>{{/isList}}{{^isList}}{{fieldClassName}}{{/isList}} {{fieldName}} =
-                modify{{#isList}}{{listClassName}}{{/isList}}{{^isList}}{{#isToken}}Token{{/isToken}}{{#isNode}}Node{{/isNode}}{{/isList}}({{camelCaseExternalClassName}}.{{fieldName}}(){{#isOptional}}.orElse(null){{/isOptional}});
-        {{/fields}}
-        return {{camelCaseExternalClassName}}.modify(
-                {{^syntaxKind}}{{camelCaseExternalClassName}}.kind(), &#10;                {{/syntaxKind}}{{#fields}}{{fieldName}}{{^isLast}}, &#10;                {{/isLast}}{{/fields}});
+        return ({{externalClassName}}) super.transform({{camelCaseExternalClassName}});
     }
     {{/isAbstract}}
     {{/nodes}}
-
-    // Tokens
-
-    @Override
-    public Token transform(Token token) {
-        return token;
-    }
-
-    @Override
-    public IdentifierToken transform(IdentifierToken identifier) {
-        return identifier;
-    }
-
-    @Override
-    protected Node transformSyntaxNode(Node node) {
-        return node;
-    }
-
-    protected <T extends Node> NodeList<T> modifyNodeList(NodeList<T> nodeList) {
-        return modifyGenericNodeList(nodeList, NodeList::new);
-    }
-
-    protected <T extends Node> SeparatedNodeList<T> modifySeparatedNodeList(SeparatedNodeList<T> nodeList) {
-        Function<NonTerminalNode, SeparatedNodeList> nodeListCreator = SeparatedNodeList::new;
-        if (nodeList.isEmpty()) {
-            return nodeList;
-        }
-
-        boolean nodeModified = false;
-        STNode[] newSTNodes = new STNode[nodeList.internalListNode.size()];
-
-        for (int index = 0; index < nodeList.size(); index++) {
-            T oldNode = nodeList.get(index);
-            T newNode = modifyNode(oldNode);
-            if (oldNode != newNode) {
-                nodeModified = true;
-            }
-
-            newSTNodes[2 * index] = newNode.internalNode();
-            if (index == nodeList.size() - 1) {
-                break;
-            }
-
-            Token oldSeperator = nodeList.getSeparator(index);
-            Token newSeperator = modifyToken(oldSeperator);
-
-            if (oldSeperator != newSeperator) {
-                nodeModified = true;
-            }
-
-            newSTNodes[(2 * index) + 1] = newSeperator.internalNode();
-        }
-
-        if (!nodeModified) {
-            return nodeList;
-        }
-
-        STNode stNodeList = STNodeFactory.createNodeList(java.util.Arrays.asList(newSTNodes));
-        return nodeListCreator.apply(stNodeList.createUnlinkedFacade());
-    }
-
-    private <T extends Node, N extends NodeList<T>> N modifyGenericNodeList(
-            N nodeList,
-            Function<NonTerminalNode, N> nodeListCreator) {
-        if (nodeList.isEmpty()) {
-            return nodeList;
-        }
-
-        boolean nodeModified = false;
-        STNode[] newSTNodes = new STNode[nodeList.size()];
-        for (int index = 0; index < nodeList.size(); index++) {
-            T oldNode = nodeList.get(index);
-            T newNode = modifyNode(oldNode);
-            if (oldNode != newNode) {
-                nodeModified = true;
-            }
-            newSTNodes[index] = newNode.internalNode();
-        }
-
-        if (!nodeModified) {
-            return nodeList;
-        }
-
-        STNode stNodeList = STNodeFactory.createNodeList(java.util.Arrays.asList(newSTNodes));
-        return nodeListCreator.apply(stNodeList.createUnlinkedFacade());
-    }
-
-    protected <T extends Token> T modifyToken(T token) {
-        if (token == null) {
-            return null;
-        }
-        // TODO
-        return (T) token.apply(this);
-    }
-
-    protected <T extends Node> T modifyNode(T node) {
-        if (node == null) {
-            return null;
-        }
-        // TODO
-        return (T) node.apply(this);
-    }
 }
-
