+++ compiler/ballerina-treegen/src/main/resources/treegen_config.properties
@@ -12,5 +12,6 @@ external.node.visitor.template=external_node_visitor_template.mustache
 external.node.transformer.template=external_node_transformer_template.mustache
 external.tree.modifier.template=external_tree_modifier_template.mustache
 external.node.factory.template=external_node_factory_template.mustache
+external.base.node.modifier.template=external_base_node_modifier_template.mustache
 external.node.package=io.ballerina.compiler.syntax.tree
 syntax.node.metadata=syntax_node_metadata.json
